# src/main/java/io/github/cuisse/countries/model/Country.java
package io.github.cuisse.countries.model;

/**
 * Represents a country and its capital.
 *
 * @param country The country.
 * @param capital The capital.
 */
public record Country(String country, String capital) {

}